requirements.txt
telethon
pillow
requests
python-dotenv